requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
python-dotenv==1.0.0
pandas==2.2.3
plotly==5.18.0
werkzeug==3.0.1
gunicorn==21.2.0